move core package to jvm-js settings
