+++ core/jvm/src/main/scala/hammock/package.scala
