+++ core/shared/src/main/scala/hammock/Codec.scala
