+++ core/shared/src/main/scala/hammock/Hammock.scala
@@ -2,15 +2,16 @@ package hammock
 
 import cats._
 import cats.arrow._
+
 import free._
-import org.apache.http.client.HttpClient
+import free.algebra._
 
 object Hammock {
 
   trait Request[A] {
     def freeReq: HttpRequestIO[HttpResponse]
-    def run[F[_] : MonadError[?[_], Throwable]](implicit httpClient: HttpClient): F[HttpResponse] = 
-      freeReq foldMap Interp.trans
+    def run[F[_] : MonadError[?[_], Throwable]](implicit interp: InterpTrans): F[HttpResponse] = 
+      freeReq foldMap interp.trans
   }
 
   class WithBodyRequest[A](val freeReq: HttpRequestIO[HttpResponse]) extends Request[A]
