+++ core/shared/src/main/scala/hammock/HttpResponse.scala
