+++ core/shared/src/main/scala/hammock/Method.scala
