+++ core/shared/src/main/scala/hammock/Status.scala
