--- core/src/main/scala/hammock/implicits.scala
@@ -1,7 +1,0 @@
-package hammock
-
-import org.apache.http.impl.client.HttpClientBuilder
-
-object implicits {
-  implicit val client = HttpClientBuilder.create().build()
-}
