+++ project/plugins.sbt
@@ -1,2 +1,3 @@
+addSbtPlugin("org.scala-js" % "sbt-scalajs" % "0.6.14")
 addSbtPlugin("com.fortysevendeg"  % "sbt-microsites" % "0.4.0")
 addSbtPlugin("me.lessis" % "bintray-sbt" % "0.3.0")
